+++ src/ast/mod.rs
@@ -11,13 +11,13 @@ mod strct;
 mod tree;
 mod ty;
 mod ty_enum;
+pub mod consts;
 
 pub use tree::*;
 // pub use imports::resolve_imports;
 pub use imports::{import_extern_symbols, imported_units_map, validate_imports};
 pub use parser::{generate_ast, parse_syntax_tree, print_fancy_parse_err};
 // pub use protos::generate_protos;
-pub use misc::import_and_sort_consts;
 pub use protos::{
     generate_protos, import_protos, rec_types_and_parse_imports_and_macros, validate_protos,
 };